Replace explicit type parameters in function calls with template parameters; Added method definitions for SparseOctree<double, uint8_t>

+++ OctreeApi/OctreeLibrary/include/types.cuh
@@ -5,7 +5,7 @@
 
 struct Averaging
 {
-    uint64_t r, g, b;
+    uint32_t r, g, b;
     uint32_t pointCount;
 };
